Move NewsAdapter into chapter package
