+++ app/src/main/java/org/gdg/frisbee/android/chapter/NewsFragment.java
@@ -34,7 +34,6 @@ import org.gdg.frisbee.android.BuildConfig;
 import org.gdg.frisbee.android.Const;
 import org.gdg.frisbee.android.R;
 import org.gdg.frisbee.android.common.GdgActivity;
-import org.gdg.frisbee.android.adapter.NewsAdapter;
 import org.gdg.frisbee.android.api.GapiOkTransport;
 import org.gdg.frisbee.android.app.App;
 import org.gdg.frisbee.android.cache.ModelCache;
